+++ src/torchpme/tuning/__init__.py
@@ -1,6 +1,6 @@
 from .ewald import tune_ewald
-from .pme import tune_pme
 from .p3m import tune_p3m
+from .pme import tune_pme
 
 __all__ = [
     "tune_ewald",
